+++ backend/routes/departments.js
@@ -4,7 +4,9 @@ const Departments = require("../models/departments");
 const Auth = require("../middleware/auth");
 const validateDepartment = require("../middleware/departments");
 
-router.post("/createDepartment", validateDepartment, async(req, res) => {
+router.post("/createDepartment", Auth, validateDepartment, async(req, res) => {
+    const user = await User.findById(req.user._id);
+    if (!user) return res.status(401).send("Error en usuario");
     const department = await Departments.findOne({department: req.body.department})
     if(department) return res.status(401).send("El departamento ya existe");
     const departments = new Departments ({
@@ -15,12 +17,16 @@ router.post("/createDepartment", validateDepartment, async(req, res) => {
     return res.status(200).send({result});
 })
 
-router.get("/getDepartments", async(req, res) => {
+router.get("/getDepartments", Auth, async(req, res) => {
+    const user = await User.findById(req.user._id);
+    if (!user) return res.status(401).send("Error en usuario");
     const departments = await Departments.find();
     return res.status(200).send({departments});
 })
 
-router.put("/editDepartment", validateDepartment, async(req, res) => {
+router.put("/editDepartment", Auth, validateDepartment, async(req, res) => {
+    const user = await User.findById(req.user._id);
+    if (!user) return res.status(401).send("Error en usuario");
     const dpt = await Departments.findByIdAndUpdate(req.body._id, {
         department: req.body.department,
         status: req.body.status
@@ -29,7 +35,9 @@ router.put("/editDepartment", validateDepartment, async(req, res) => {
     return res.status(200).send({dpt})
 })
 
-router.delete("/:_id", async(req, res) => {
+router.delete("/:_id", Auth, async(req, res) => {
+    const user = await User.findById(req.user._id);
+    if (!user) return res.status(401).send("Error en usuario");
     const department = await Departments.findByIdAndDelete(req.params._id);
     if (!department) return res.status(401).send("No existe el departamento");
     return res.status(200).send("Departamento eliminado");
